Unconditionally require pyFFTW in Cython kernel.
Conditional compilation "IF" directives are deprecated in the latest
Cython. We should also be able to install pyFFTW as part of the build
without too many issues. The simplest thing is likely to just require
it in all builds. It may be possible to make pyFFTW optional again in
the future.